requirements.txt
google-cloud-storage==1.16.1
numpy==1.23.0
pandas==1.5.0
tld==0.12.5
joblib==1.1.1

